правки файла сборки

+++ src/rest_compare.cpp
@@ -142,7 +142,7 @@ void compare_get(http_request request)
 
                 for (size_t i = 0; i < 2; i++) free(tmps[i]);
 
-                answer[ELEMENT_VALUE] = json::value::number((double)score);
+                answer[ELEMENT_VALUE] = json::value::number(score);
                 answer[ELEMENT_RESULT] = json::value::boolean(true);
                 answer[ELEMENT_TYPE] = json::value::string(conversions::to_string_t(to_string(compare_type)));
             }
